third_party/WebKit/LayoutTests/platform/win/external/wpt/IndexedDB/idbdatabase_deleteObjectStore-expected.txt

This is a testharness.js-based test.
PASS IDBDatabase.deleteObjectStore() - object store's name is removed from database's list  
Harness: the test ran to completion.
